Fix one undef behavior.

+++ lib/ccv_output.c
@@ -7,11 +7,11 @@ int ccv_cli_output_level_and_above(int level)
 	if (level == CCV_CLI_NONE)
 		return CCV_CLI_NONE;
 	int i;
-	int levels = 0;
+	uint32_t levels = 0;
 	for (i = 0; i < 32; i++)
-		if (level <= (1 << i))
-			levels |= 1 << i;
-	return levels;
+		if (level <= (1u << i))
+			levels |= 1u << i;
+	return (int)levels;
 }
 
 void ccv_set_cli_output_levels(int levels)
